fixed src/ added touchable w/o feedback removed api calls
