+++ client/App.js
@@ -2,7 +2,7 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View, Image, SafeAreaView} from 'react-native';
 
-import Navigator from './screens/Navigation/HomeStack'
+import Navigator from './src/Navigation/HomeStack'
 import { NavigationContainer } from '@react-navigation/native';
 
 export default function App() {
